capitalize Jsonnet in some places
Just the bare minimum add new text to explain the change to kubecfg

This PR should be reopened as the `jsonnet-gutted` branch, with
commentary removed and slimmed down to just code examples with as little
non-embedded commentary as possible.

This way we can consider reorganizing any pieces that would need to be
redone for example in a hypothetical Jenkins+cdk8s use case document, or
something else like Gitlab+Tanka, or Tekton+kubecfg, or any combination.

If we are going to document outside solutions and proper integrations
with them, then we should probably make their treatments as equal as
possible to avoid showing favoritism.