Update bootstrap github docs to list permissions
Document the required permissions for bootstrapping pre-existing repositories.

I'm not sure if this is the correct place to document this as it feels to detailed for CLI output
but (as I understand it) the output of this is then passed to the website automatically and, not
knowing this took a while to debug the problems I was having.

Signed-off-by: Sam Cook <sam.cook@weave.works>